Initial commit - incl. NCurses
Proof of concept, but recursive updating is of course computationally
impossible as GHC is not automatically momoizing the tick function.

Note: added NCurses sources as modification was needed to have NCurses build
on OS X Sierra (issue with language-c package - used by c2hs - which is
unable to parse some Apple extensions in headers of stdio.h.